Finished /tck-delete
The command tck-delete was finished by finishing its run function. Also the function deleteTicket in utilis/ticketManager was finished in order to be used in tck-delete run funtion.
Was also edited the function getTicketChannel in utilis/ticketManager to include the case of a closed ticket and to make it more readable.