Added command /tck-reopen
Added the command /tck-reopen for ticketing along with the function reopenTicket in utilis/ticketManager used in the run function of the command. Added the command data in language.yml and commands.yml.
Was also added an error in errors.yml to be triggered when trying to reopen an unclosed ticket.