General code reorganisation
The command building was decentralized to structures/Command.js instead of being locally did by each command class. To do so the command options data stored in language.yml was moved to commands.yml.
Was also solved the problem wich caused the application to be unresponsive due to a too long command run time by deferring the reply to the interaction immediatly after the trigger.